merged to upstream/master

+++ code/modules/admin/buildmodes/spawn_single.dm
@@ -6,6 +6,7 @@ Ctrl-RMB on buildmode button = Set cinematic effect<br>
 
 Left Mouse Button on turf/mob/obj      = Place objects<br>
 Right Mouse Button                     = Delete objects<br>
+Right Mouse Button + Shift             = Set object type to selected mob/obj type<br>
 <br>
 Use the button in the upper left corner to<br>
 change the direction of created objects.<br>
@@ -16,7 +17,7 @@ change the direction of created objects.<br>
 	var/matrix/mtx = matrix()
 	click_mode_right(var/ctrl, var/alt, var/shift)
 		if(ctrl)
-			cinematic = (input("Cinematic spawn mode") as null|anything in list("Telepad", "Blink", "Supplydrop", "Supplydrop (no lootbox)", "None")) || cinematic
+			cinematic = (input("Cinematic spawn mode") as null|anything in list("Telepad", "Blink", "Supplydrop", "Supplydrop (no lootbox)", "Lethal Supplydrop", "Lethal Supplydrop (no lootbox)", "None")) || cinematic
 			return
 		objpath = get_one_match(input("Type path", "Type path", "/obj/closet"), /atom)
 		update_button_text(objpath)
@@ -51,7 +52,7 @@ change the direction of created objects.<br>
 
 
 						if (isobj(A) || ismob(A) || isturf(A))
-							A.dir = holder.dir
+							A.set_dir(holder.dir)
 							A.onVarChanged("dir", SOUTH, A.dir)
 						sleep(0.5 SECONDS)
 						mtx.Reset()
@@ -70,15 +71,37 @@ change the direction of created objects.<br>
 						A = new objpath(T)
 
 					if (isobj(A) || ismob(A) || isturf(A))
-						A.dir = holder.dir
+						A.set_dir(holder.dir)
 						A.onVarChanged("dir", SOUTH, A.dir)
 						blink(T)
 				if("Supplydrop")
 					if (ispath(objpath, /atom/movable))
 						new/obj/effect/supplymarker/safe(T, 3 SECONDS, objpath)
+					else if(ispath(objpath, /turf))
+						T.ReplaceWith(objpath, handle_air = 0)
+					else
+						new objpath(T)
 				if("Supplydrop (no lootbox)")
 					if (ispath(objpath, /atom/movable))
 						new/obj/effect/supplymarker/safe(T, 3 SECONDS, objpath, TRUE)
+					else if(ispath(objpath, /turf))
+						T.ReplaceWith(objpath, handle_air = 0)
+					else
+						new objpath(T)
+				if("Lethal Supplydrop")
+					if (ispath(objpath, /atom/movable))
+						new/obj/effect/supplymarker(T, 3 SECONDS, objpath)
+					else if(ispath(objpath, /turf))
+						T.ReplaceWith(objpath, handle_air = 0)
+					else
+						new objpath(T)
+				if("Lethal Supplydrop (no lootbox)")
+					if (ispath(objpath, /atom/movable))
+						new/obj/effect/supplymarker(T, 3 SECONDS, objpath, TRUE)
+					else if(ispath(objpath, /turf))
+						T.ReplaceWith(objpath, handle_air = 0)
+					else
+						new objpath(T)
 				else
 					var/atom/A = 0
 					if(ispath(objpath, /turf))
@@ -87,9 +110,14 @@ change the direction of created objects.<br>
 						A = new objpath(T)
 
 					if (isobj(A) || ismob(A) || isturf(A))
-						A.dir = holder.dir
+						A.set_dir(holder.dir)
 						A.onVarChanged("dir", SOUTH, A.dir)
 
 	click_right(atom/object, var/ctrl, var/alt, var/shift)
-		if(isobj(object))
-			qdel(object)
+		if (shift)
+			if (ismob(object) || isobj(object))
+				objpath = object.type
+				update_button_text(objpath)
+		else
+			if(isobj(object))
+				qdel(object)
